Implement Autohide for Mobile/Touch Devices.
A touch on the videoplayer fades in the player controls, over 'fadeIn' milliseconds.
When the touch is released, after 'hold' milliseconds, the controls fade out after 'fadeOut' milliseconds,
unless the player or controls are retouched, in which case the 'hold' countdown resets until next touch release.

I copied the conditional checks for autohide into the event handlers.
This satisfies the need to not autohide the controls on video end and on video pause, for example.
Many video players, particularly for entertainment content, keep the controls shown on end and on pause.
I have not tested fully if having the conditionals here negates the need for the same conditional on line 2262.

I added an option, nativeControlsHideOverride, defaulted to false.